+++ routes/Inventory/Licence.js
@@ -2,29 +2,39 @@ const express = require('express')
 const router = express.Router()
 const {body, validationResult} = require("express-validator");
 const Licence = require("../../models/Invenory/Licence");
+const {checkTokenDate} = require("../../functions/token");
 
 /*Portal Licence add new item*/
 router.post('/',
     async (req, res)=>{
         try {
-            const errors = validationResult(req)
-            const {org, seller, vendor, lic, key, start, exp, info, notes, amount, status,  _id} = req.body
-
-            if (!errors.isEmpty()){
-                return res.status(400).json({success: false, errors: errors.array()})
+            const authHeader = req.headers['authorization'];
+            if (!authHeader) {
+                return res.status(401).send('Authorization header is missing');
             }
-            let licItem;
+            const bearer = authHeader.split(' ')[1];
+            const token = await checkTokenDate(bearer)
+            if (token.token){
+                const errors = validationResult(req)
+                const {org, seller, vendor, lic, key, start, exp, info, notes, amount, status,  _id} = req.body
 
-            if (_id){
-                await Licence.replaceOne({"_id": _id}, {
-                    'org': org, 'seller': seller, 'vendor': vendor, 'lic': lic, 'key': key,'start': start,'exp': exp,'info': info,'notes': notes,'amount': amount,'status':status,
-                })
-            } else {
-                licItem = new Licence({org, seller, vendor, lic, key, start, exp, info, notes, amount, status: true})
-                await licItem.save()
+                if (!errors.isEmpty()){
+                    return res.status(400).json({success: false, errors: errors.array()})
+                }
+                let licItem;
+
+                if (_id){
+                    await Licence.replaceOne({"_id": _id}, {
+                        'org': org, 'seller': seller, 'vendor': vendor, 'lic': lic, 'key': key,'start': start,'exp': exp,'info': info,'notes': notes,'amount': amount,'status':status,
+                    })
+                } else {
+                    licItem = new Licence({org, seller, vendor, lic, key, start, exp, info, notes, amount, status: true})
+                    await licItem.save()
+                }
+                const result = {id: 200,message: "Запись успешно добавлена",}
+                res.json({result})
             }
-            const result = {id: 200,message: "Запись успешно добавлена",}
-            res.json({result})
+
         }catch (e) {
             res.status(500).send({message: e.message})
         }
@@ -33,8 +43,17 @@ router.post('/',
 
 router.get('/', async (req, res) => {
     try{
-        const collection = await Licence.find({})
-        res.json(collection)
+        const authHeader = req.headers['authorization'];
+        if (!authHeader) {
+            return res.status(401).send('Authorization header is missing');
+        }
+        const bearer = authHeader.split(' ')[1];
+        const token = await checkTokenDate(bearer)
+        if (token.token){
+            const collection = await Licence.find({})
+            res.json(collection)
+        }
+
     }catch (e) {
         console.log(e)
     }
@@ -42,11 +61,20 @@ router.get('/', async (req, res) => {
 
 
 router.delete('/', async (req, res) =>{
-    console.log(req.body.id)
+   /* console.log(req.body.id)*/
     try{
-        await Licence.deleteOne(req.body.id)
-        const result = {id: 200,message: "Номер успешно удален"}
-        res.json(result)
+        const authHeader = req.headers['authorization'];
+        if (!authHeader) {
+            return res.status(401).send('Authorization header is missing');
+        }
+        const bearer = authHeader.split(' ')[1];
+        const token = await checkTokenDate(bearer)
+        if (token.token){
+            await Licence.deleteOne(req.body.id)
+            const result = {id: 200,message: "Номер успешно удален"}
+            res.json(result)
+        }
+
     }catch (e) {
         console.log(e)
     }
